backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai==1.61.1
jsonschema==4.23.0
redis==5.2.1
rq==2.0.0
pillow==11.1.0
pytest==8.3.4
httpx==0.28.1
